requirements.txt
absl-py==1.2.0
altair==4.2.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
attrs==22.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.5
cachetools==5.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
commonmark==0.9.1
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
entrypoints @ file:///C:/ci/entrypoints_1649926621128/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
flatbuffers==2.0.7
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
ipykernel @ file:///C:/b/abs_21ykzkm7y_/croots/recipe/ipykernel_1662361803478/work
ipython @ file:///C:/ci/ipython_1657634415474/work
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.1.2
jsonschema==4.16.0
jupyter-core @ file:///C:/ci/jupyter_core_1651656286743/work
jupyter_client @ file:///C:/b/abs_8fbm7986b_/croots/recipe/jupyter_client_1662504374117/work
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929390/work
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.4
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
pyrsistent==0.18.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pyzmq @ file:///C:/ci/pyzmq_1657615952984/work
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
semver==2.13.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
streamlit==1.12.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow_intel==2.10.0
termcolor==2.0.1
toml==0.10.2
toolz==0.12.0
tornado @ file:///C:/ci/tornado_1662458743919/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
